+++ Solid.Data/Repositories/BorrowRepository.cs
@@ -6,6 +6,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Library;
 using Library.Entities;
+using Microsoft.EntityFrameworkCore;
 
 namespace Solid.Data.Repositories
 {
@@ -14,8 +15,9 @@ namespace Solid.Data.Repositories
         private readonly DataContext _context;
         public List<Borrow> GetBorrows()
         {
-            return _context.Borrows.ToList();
+            return _context.Borrows.Include(b => b.Member).Include(b => b.Books).ToList();
         }
+
         public Borrow GetById(int id)
         {
             return _context.Borrows.Find(id);
@@ -32,10 +34,14 @@ namespace Solid.Data.Repositories
         public Borrow Put(int id, Borrow value)
         {
             Borrow b = _context.Borrows.Find(id);
-            //b.Status = value.Status;
-            //b.Author = value.Author;
-            //b.Title = value.Title;
-            //b.Description = value.Description;
+            if (b != null)
+            {
+                b.Member = value.Member;
+                b.MemberId = value.Member.Id;
+                b.Books = value.Books;
+                b.Date = value.Date;
+                b.Status = value.Status;
+            }
             _context.SaveChanges();
             return b;
 
@@ -53,7 +59,8 @@ namespace Solid.Data.Repositories
         public Borrow Delete(int id)
         {
             Borrow b = _context.Borrows.Find(id);
-            _context.Borrows.Remove(b);
+            if (b != null)
+                _context.Borrows.Remove(b);
             _context.SaveChanges();
             return b;
         }
